Initial commit
Use DSharpplus as the main library for this bot.

Add in various elements from kingbot which includes the CredentialsHelper,
DI syntax, and CommandHandler.

Since D#plus handles messages in a different way, I had to create
a new string renderer class which contains methods to convert the
message content into a string which we can manipulate.

For now, the StringRenderer class is synchronous since there's no
reason to use async due to lack of await statements.

Add a ping command to see if the bot is online and add a ban command
which allows to user to ban with a reason that's printed when
the user is successfully banned.

Signed-off-by: kingbri <bdashore3@gmail.com>