Introduce a warning system
This system allows for administrators to warn users. Warns are
usually given before a kick or ban to make sure the user complies
with the rules. After a certain amount of warnings (usually 3), the
user is banned.

This warning system is configurable for how many warns are given until
the user is banned through modification of the info.json file (Will be changed).

To use the warns system, I brought up code for a postgres database
which stores the user's ID and the amount of warns the user has
recieved.

You can also remove a warn from a user by a seperate command which
removes ONE warn. If the warned user's count becomes 0, the entry is
removed from the database.

Finally, there's one more command which shows how many warns a user has.

Additional housekeeping:
- Added an unban method
- Set up framework for mutes
- Added a try/catch method to log all errors

Signed-off-by: kingbri <bdashore3@gmail.com>