Add a temporary permission system
Add a PermissionHandler to check if the user is a guild admin,
this is used for multiple instances such as warns, bans, and
mutes (Planned for the future).

Added a temporary permissions check on admin only commands (bans, warns).
This is redundant code and is expected to be removed in a future commit
once I figure out a proper admin system.

Also add a template class for muting/unmuting a user.

Signed-off-by: kingbri <bdashore3@gmail.com>