Add a system for welcome/leave messages
I wanted to make this as easy as possible, so here's how it works

- The database entry is created by a command to store your guild
in the db for welcome/leaves.

- The user can customize the welcome and leave messages to their
liking. He/she can also change the channel where the message is sent.
The channel the user typed the init command is default.

- Users can use variables {member}, {user}, and {server} to dynamically
change the welcome message per user and server. For channel names, the
user should use normal # channel syntax.

- Users can purge the guild from the database if they found another bot
for welcoming users.

- Leave messages and welcome messages can be individually set. If any
of them are not set or cleared, the bot will not do anything when a
new member joins.

All welcome/leave message configuration is restricted to server admins.

Signed-off-by: kingbri <bdashore3@gmail.com>