Add in embeds, a help command, and custom prefixes
Embeds: Most messages sent by the bot will usually have embeds to
make them look cleaner and more like an actual admin bot. The embeds
for module classes are stored within a common EmbedStore class. Help
embeds will be either stored in their respective modules or a seperate
class (Haven't decided yet).

Help Command: This help command is modular in nature where there is a
base help command which grabs all the help subcommands for the user.

Custom Prefixes: Every time a message is sent, a new PrefixHelper is
created which checks if the prefix is the same as the server's set
prefix. If there is no prefix row within the database, the function
returns the default prefix set in CredentialsHelper.

Custom prefixes are possible due to EF Core's first level caching
when pulling something from a database which improves speed and overall
responsiveness.

Signed-off-by: kingbri <bdashore3@gmail.com>