Add mute/unmute support
This commits adds support for muting and unmuting users within a guild.

The bot creates a mute role and sets mute permissions in all channels.
The mute role ID is stored inside the database, so we can assign the user
the mute role when the mute command is called.

The commands also check if a user is already muted.

Signed-off-by: kingbri <bdashore3@gmail.com>