Modify prefixes to use a string trie
Instead of using direct database calls (which can be extremely slow),
use a middle I/O layer which allows for fast storage and retrieval of
the prefixes.

When adding/updating a new prefix, the prefix is both added in the
trie and the database at the SAME time due to asynchronous operations
regarding the database. All operations are done from the trie, the
database is a backup.

When the bot is restarted (from a crash or something), all prefixes
are automatically read into the trie, so no data is lost.

The only problem here was the service lifespan of the PrefixHelper which
was Transient. This is not okay because every time we tried to read
the prefix, GC deleted every class instantiation due to DI's nature.
The nature of a trie basically requires us to have a single instance
of PrefixHelper and changing its lifespan to Scoped fixed the issue.

Suggested-by: Tanish Manku <tanish2k09.dev@gmail.com>
Signed-off-by: kingbri <bdashore3@gmail.com>