Add time-based mutes and helpful comments
Huge thanks for tanish2k09 for creating the base TrieHandler
which manages all aspects of the timer cache for mutes. He
is co-authored on this commit for that reason.

Changes:
1. Add an extra cache layer for prefixes and time-based mutes
using a stringtrie that takes a string and a class. This class
has the prefix and mute time trie within it. This system allows
us to have a 3 parameter system instead of two.

2. Using this cache layer, time-based mutes now exist within RoyalGuard.
The admin mutes the mentioned user for either seconds, minutes, hours,
days, or weeks. The given time is converted into ms for the timers.

The timer is stopped when:
- It expires
- An admin manually unmutes the user

In both cases, an unmute embed is sent.

TODO: Add a time value in the mute embed so the user can see
how long he/she is muted for.

Co-Authored-by: kingbri <bdashore3@gmail.com>
Co-Authored-by: Tanish Manku <tanish2k09.dev@gmail.com>