Hotfix: Fix mute timers not being added
If the guild isn't in the trie, add it because the users are using
the default prefix, add the guild into the trie with the default prefix.

This is accomplished by converting the AddNewMute function into a bool.
If the bool returns false, the function is re-ran since a trie entry
was added.

This causes the timer to properly stop and not keep running and error
out due to the database not finding the query for the second time.

Signed-off-by: kingbri <bdashore3@gmail.com>