Various updates and hotfixes
hotfix: Make the bot ignore messages for itself. This was mainly
my stupidity for not including this as the first thing to do.

Update the commands to set the channel where timed unmutes are sent.
This is a better implementation than storing timed mute channels per
mute because it makes timed unmuting cleaner if a user has been
muted for a long time.

Signed-off-by: kingbri <bdashore3@gmail.com>