botwide: Add an emergency system to recover from mistakes
If an admin mistakenly sets the prefix to that of another bot, the
two bots will get confused. Implement an emergency system for the
admin to quickly reset and/or update the prefix using "eg"command

This emergency system has an admin check to make sure users don't
abuse this command and the bot says an emergency command has been sent.

Signed-off-by: kingbri <bdashore3@gmail.com>