Improve emergency check to use mentions
Using a bot mention is better for an emergency command since
someone can set the bot prefix to use "eg". Using a mention is unique
and can't be replicated. (As suggested by Tanish)

Also, remove the split between unban and unbanId commands. The unban
command works with both Ids and mentions.

Suggested-by: Tanish Manku <tanish2k09.dev@gmail.com>
Signed-off-by: kingbri <bdashore3@gmail.com>