Merge prefixes with guild info storage
Having a seperate table for a custom prefix is not needed because
we're only storing one value

The DB has been updated with the migrations and no data has been lost.

Also add a GuildInfoHelper for ensuring that a db entry exists and
adding new entries. All updates are handled in their respective
classes because those functions are unique per command

Signed-off-by: kingbri <bdashore3@gmail.com>