Add a orphan guild collection system
When someone kicks the bot, the guild data still persists in the database.

Convert the bot to use hosted services instead of fetching the bot's
service from a container. This allows for a background service which
polls the GuildInfoStore table every 4 days to check if orphaned
guild data needs to be deleted.

When the admin kicks the bot from the server, a timestamp + 1 week
is added into the database. The collector checks whether the stored
time is less than or equal to the current time. If so, the data is
deleted from the DB using a delete cascade.

If the admin re-adds the bot before that week expiry period, the delete
timestamp is reset back to 0 and the data is preserved.

Signed-off-by: kingbri <bdashore3@gmail.com>