Add support to give roles on welcome event
When a user joins the server, roles can now be automagically given.
This is NOT the same as reaction roles where the user clicks a reaction
and a role is assigned. Reaction roles are scheduled for v2.0.

Admins can configure the role Ids to be stored in the database as long
as these roles are mentionable. If a role already exists in the database,
the bot will let the admin know in the embed itself.

In addition, the admin can add/remove multiple roles in one command.

If the role is removed from the server, the bot will automatically
delete the database entry when the next user joins to increase efficiency
and reduce un-needed storage space.

Welcome roles are a subcommand of the "welcome" command and not their own
command!

Signed-off-by: kingbri <bdashore3@gmail.com>