Add a kick command and various cleanup
Upgrade the bot to use a ShardedClient instead of the basic DiscordClient.
This required a modification of guild fetching inside the mutes class.

Also, add a kick command in the Other class since it doesn't need
its own class.

Finally, modify all mentions to use the D#plus Mention string
instead of creating it manually for most cases unless an ID is used
for mentioning a user.

Signed-off-by: kingbri <bdashore3@gmail.com>