Add a moderator permissions layer
Moderators have to have the "ban users" permission for RoyalGuard
to recognize them as such.

It has come to my attention that there are not two types of server
permissions, but there are three, Admin, Moderator, and User.

Modify commands to differentiate between moderator and admin usage
since some cannot be trusted for mods.

Signed-off-by: kingbri <bdashore3@gmail.com>