Use a bolded username/discriminator when a user leaves
If the user leaves the server, the message shows invalid-user on
the phone which is a bug in the discord client and can't be circumvented
when using a mention.

Therefore, make the user leave event show a bolded username/discriminator
combo so the admins can track down who left. This is not an elegant solution,
but it's the only one that could be done.

Signed-off-by: kingbri <bdashore3@gmail.com>