Introduce dynamic prefixes and permissions
the permissions_helper catches any invalid permissions that are
in a certain case of a command rather than using a dispatch error
to lock someone out of a command if the user doesn't have the permission.

Dynamic prefixes are also present in this bot.

The guild_info table is altered when a new prefix is set. A row
in guild_info is created when a bot enters a guild and will eventually
persist for a week when the bot is kicked out.

All tables below guild_info will have a cascade deletion meaning that
when the guild_info row is deleted, all rows with the guild_id key will
be removed as well.

Signed-off-by: kingbri <bdashore3@gmail.com>