Add configurable moderation roles
RoyalGuard now has a new permission system based on moderation roles
rather than checking for a static permission.

The server administrator has to configure the moderator role using
the `moderator` command. This command makes sure the role has permissions
`Ban Members` and `Manage Messages` as they are both required for
the bot's functionality.

More required permissions will be added as more features are written.

The bot first looks for if the user's an administrator. If the user
isn't an admin, proceed with the mod role check.

Improve bans with this functionality and add some error handling.

Resolves #7

Signed-off-by: kingbri <bdashore3@gmail.com>