Implement warnings
Warns are a step behind bans in a server.

There are 3 warnings before the user gets banned. The given punishment
and number of warns to ban will be configurable past v2.0.

An admin can warn/unwarn using the respective commands and any
user can get warns of self or a mention.

Signed-off-by: kingbri <bdashore3@gmail.com>