Add mutes
Mutes silence a user either for a period of time or permanently. This
is accomplished by inferring if the user is inputting a time or a normal
mention-based mute.

If a mute is timed, spawn a new task that delays a future until
completion. The future can be aborted at any time by manually unmuting
the user.

If the future is aborted, the database row is quietly deleted. If it
can't be deleted, the next timed mute will overwrite the existing one
and the delete error will be logged in the console.

Otherwise, unmute the user using the mute channel. The mute channel
is created when the "muted" role is created within the server. This
channel can be changed using the `mutechannel` command at any time by
moderators.

Signed-off-by: kingbri <bdashore3@gmail.com>