Cache prefixes in a Dashmap
Creating a write-through cache for prefixes. This was needed because
some commands require the length of the bot's prefix in order
to get a specific word.

Adding this cache also reduces the amount of sql calls per command.
It's easier to fetch a prefix from memory rather than an external database.

To make it more optimized, only store the guilds that have custom prefixes
in the Dashmap. If there is no entry for the guild inside the map, the
dynamic_prefix function returns None. Doing this forces the bot to use
the default prefix.

Signed-off-by: kingbri <bdashore3@gmail.com>