Add kicks and modify permissions
Add the ability to kick a user with a reason. Also, edit how reasons work
by storing the user who created the reason.

Also, make permissions not return Err since the new error system
sends an error message when that happens. Instead, let the user know
what to do to stop the errors.

To combat this, also add a clear command for the moderator role which
forces the bot to avoid the mod role database fetch and doesn't send
any errors.

In the future, I'm going to add an embed which shows the current bot
configuration.

Signed-off-by: kingbri <bdashore3@gmail.com>