Add a delete buffer system
If a user kicks/bans the bot from the guild, the data persists for a
week before getting lost.

This prevents any accidental deletion of data and saves the pain of the
admin reconfiguring the entire bot. Instead, all the data is in place
for a week.

After a week, if the admin decides to reinvite the bot, the data is lost.

Signed-off-by: kingbri <bdashore3@gmail.com>