Add emergency support
If the user can't reconfigure the bot's prefix since there's a conflicting
bot with the same command, the user can mention the bot to do so.

This opens up a realm called emergency commands. These are used if
there's something wrong with the bot's configuration and can be easily
fixed with a simple mention.

Emergency commands are either the same as the normal commands or they're
unique to the situation.

Signed-off-by: kingbri <bdashore3@gmail.com>