Add reaction role commands
There are 3 commands that have been added

1. New:
  This command is for people who have experience with creating reaction
  roles and don't want to run the wizard for each one. It's one command
  that creates one role/emoji combo.
2. Remove:
  This command removes the reaction role given the message ID, the channel
  the message is in, and the emoji used for the role.
3. Wizard:
  This is a step-by-step process to add a new reaction role. It times out
  on each step after 2 minutes so the user isn't locked inside the collector cascade.
  If the role/emoji combo already exists, the wizard exits with an error at the end.

These commands support custom emojis both from the guild and all other servers the
bot is in assuming that the user has discord nitro.

RoyalGuard's reaction roles are a simplified version of the other alternatives.
There are no reaction embeds in the current plan for the bot.

Based on my experience, I've found that plain and established messages are cleaner.

Signed-off-by: kingbri <bdashore3@gmail.com>