Add reaction role lists and update formatting
Reaction role lists are formatted in an embed. If a role/channel mention
is invalid, the guild owner is advised to remove it using `rr remove`.

There will be an automatic way to sanitize reaction roles in the future using
the all reactions cleared event.

Make the help strings use concat!() instead of insertion into a mutable string.

Signed-off-by: kingbri <bdashore3@gmail.com>