Add a reaction role leftover collection system
When a mod/admin deletes a message with reaction roles on them,
the role entries stay within the database.

To combat this, add hooks for message deletes and when all reactions
are cleared. Once an event happens, the database is checked to see
if the row exists, and if that's the case, delete the entry.

The reason I'm using exists instead of one sql delete query is because
if a row doesn't exist with said message ID, the query immediately bails
instead of running a delete query to full execution.

Signed-off-by: kingbri <bdashore3@gmail.com>