Permissions: don't assume user is in guild
When executing a command using an ID instead of a mention, the bot
would error out when checking if a user is a moderator. This is
because it checked if a user had a certain guild role.

Now, if there's an error when unwrapping the role, the function
returns false since the user doesn't exist in the guild.

Signed-off-by: kingbri <bdashore3@gmail.com>