Fix the error handling of mute deletions
When the bot starts up, it checks the database to see if a user needs to
be unmuted. However, with the introduction of the delete GC, this
caused the bot to error since it's no longer in the guild.

Fix this by checking the delete table before doing anything. If
the guild id isn't in the delete table, proceed with the mute deletion
otherwise skip.

However, if there's a rogue database value, the bot would fire an error
when trying to unwrap the cached guild. Fix this by adding a simple error
handler to let the dev know that there was a guild deletion error.

From there, the bad entry can be manually removed.

Resolves #14

Signed-off-by: kingbri <bdashore3@gmail.com>