Delete cached prefixes on guild delete
If a guild removes the bot and the delete buffer decides to remove
the guild from the db, the cached prefix will persist until the bot
is restarted from the server-side.

There's no use holding a prefix for a guild that permanently removed
the bot, so delete that as well.

Also, take the data locks outside of the main guild delete loop
since it's more efficient to do so.

Signed-off-by: kingbri <bdashore3@gmail.com>