Add a guild pruner
Sometimes the bot may be offline and leftover guilds will be stored
in the database. This introduces a GC that's run on bot startup which
checks for 2 conditions

1. If the guild in question is present within the ctx cache
2. If the guild in question is present within the delete buffer.

If the guild is in the delete buffer, the data is going to be deleted
anyways and the guild still has a chance to recover it.

Resolves #17

Signed-off-by: kingbri <bdashore3@gmail.com>