Reaction roles: Fix more than one character issue
Some unicode emojis consist of > 1 character. Since there's > 1 character
for these emojis, we have to check if each character in the string
chars is an emoji, emoji component, keycap, etc.

Doing this generates infinite possibilities of error, so instead
match on the error generated when trying to react to a message with
the given emoji.

Therefore, there is no possibility of error when creating a reaction
role.

Resolves #18

Signed-off-by: kingbri <bdashore3@gmail.com>