Fix mod role bug and update dependencies
There was a bug where the mod role was completely overwriting the database
every time the role was set.

This caused any commands executed by set moderators to throw an error.

The server's database has been completely cleared, so moderator
roles will need to be reset.

Also update sqlx to v0.5.1 and serenity to latest git head on current.

Signed-off-by: kingbri <bdashore3@gmail.com>