Fix preemptive banning
The permissions system had a bug where if a user isn't in a server,
we can't get the permissions due to not being in the server and therefore
the member cache.

Add a match case when fetching the user info. If it errors, the user
isn't an admin or mod, so break out of the check as false.

This case was present in the admin check, but not the moderator one.

Signed-off-by: kingbri <bdashore3@gmail.com>