Mutes: Refine unmuting by time
If a user gets banned from a server, the guild's members won't contain
this user.

Silently exit out on the mute check on bot start if this is the case.

Also add an extra handler if a guild member's role can't be removed
and notify the user to remove it manually since the entry is deleted
from the database before anything else happens.

Signed-off-by: kingbri <bdashore3@gmail.com>