Warns/Mutes: Fix how user object is fetched
There's no need to check mentions on a user ID error since a mention
is already parsed into a user ID and will always return Ok.

Instead, add two distinct errors to check for a user ID on arg parsing
and a validation error if the user ID isn't a valid user.

Signed-off-by: kingbri <bdashore3@gmail.com>