RoyalGuard: Add message logging
This is a non-invasive way of logging messages which uses the
serenity cache to temporarily store messages. When someone
edits or deletes a message, put a message in the designated logging
channel.

The logging channel can be enabled and disabled on a command. The
support is currently for messages, but will be expanded for server
and mod actions in the future.

Signed-off-by: kingbri <bdashore3@gmail.com>